requirements.txt
click==8.1.7
hydra-core==1.3.2
matplotlib==3.7.2
pytorch_lightning==2.1.3
scikit_learn==1.3.0
torch==2.1.0
